update pr with at least one test
